feat: Add centered buttons and self-contained table scrolling
This commit introduces changes to improve the layout and functionality of the HTML template and SCSS stylesheets. The following modifications were made:

Added the is-centered class to the buttons in the "Operating Systems," "Architectures," and "Stable Versions" sections to keep them centered even when they overflow onto a new line.
Wrapped the table in a <div> element with the class table-container to enable horizontal scrolling when the table content exceeds the available width on mobile devices.
Applied CSS styles to the table-container and table classes to achieve the desired self-contained scrolling behavior.
These changes enhance the user experience by ensuring the buttons remain centered and enabling independent horizontal scrolling for the table on mobile devices.

Modified files:

cmd/update-index/data/index.html.template
src/mystyles.scss

Signed-off-by: rsadowsk <richard.j.sadowski@gmail.com>